Refactored Helper function assert_paginated_response
