+++ backend/app/tests/api/routes/test_tag.py
@@ -7,10 +7,9 @@ from fastapi.testclient import TestClient
 from app.api.deps import SessionDep
 from app.core.config import settings
 from app.models import Organization, Tag, TagType, User
-from app.tests.utils.organization import assert_paginated_response
 from app.tests.utils.user import create_random_user, get_current_user_data
 
-from ...utils.utils import random_lower_string
+from ...utils.utils import assert_paginated_response, random_lower_string
 
 
 def setup_user_organization(
